requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
requests>=2.28
scipy>=1.10
statsmodels>=0.14
python-okx>=0.3
filelock>=3.12
